aggiornamento salvataggio locale

+++ Backend/services/sbom_parser.py
@@ -340,7 +340,7 @@ def get_docker_analysis(dockerfile_content, build_context):
         previous_filesystem = None
         filesystem_path = None
         # per test solo i primi 8 step
-        #steps = steps[:7]
+        steps = steps[:8]
         for step in steps:
 
             # Costruzione immagine dello step
@@ -388,10 +388,14 @@ def get_docker_analysis(dockerfile_content, build_context):
                             new_files
                         )
                 finally:
-                    shutil.rmtree(
-                        os.path.dirname(filesystem_path),
-                        ignore_errors=True
-                    )
+                                    
+                    previous_filesystem = filesystem_path
+                    
+                    if filesystem_path:
+                        shutil.rmtree(
+                            os.path.dirname(filesystem_path),
+                            ignore_errors=True
+                        )
 
                 step.yara_results = step_yara_results
 
@@ -468,7 +472,7 @@ def get_docker_analysis(dockerfile_content, build_context):
     finally:
         if filesystem_path:
             shutil.rmtree(
-                filesystem_path,
+                os.path.dirname(filesystem_path),
                 ignore_errors=True
             )
         builder.cleanup()
